src/raspberrryPi/requirements.txt
paho_mqtt==1.2.2
sense_hat==2.2.0